+++ TaskLogger.js
@@ -27,21 +27,21 @@ class TaskLogger {
    *    b) tag: uuid
    *    c) message: endMessage
    * 5. Returns the value from step 2
-   * @param {Object} logger
-   * @param {Function} task A function to invoke. One argument is passed to this function: an object with the properties
+   * @param {object} logger
+   * @param {function} task A function to invoke. One argument is passed to this function: an object with the properties
    * 'logger' and 'uuid.'
    * @param {*} beginMessage A message to be logged before invoking the task
    * @param {*} endMessage A message to log when the task does not throw an exception
    * @param {*} [errorMessage] A message to log when the task throws an exception. errorMessage can be overridden by the
    * provided errorHandler.
-   * @param {Function} [errorHandler] A function that is invoked with the following arguments when the task throws an
+   * @param {function} [errorHandler] A function that is invoked with the following arguments when the task throws an
    * exception:
    * 1) {Error} The exception thrown by the task
-   * 2) {Object} The logger argument
+   * 2) {object} The logger argument
    * 3) {*} The errorMessage argument
    * The function returns either the message to log or a falsey value indicating nothing should be logged. The exception
    * is rethrown regardless of the return value.
-   * @return {Promise} Resolves to the value returned or the exception thrown by the task
+   * @returns {Promise} Resolves to the value returned or rejects using exception thrown by the task
    */
   static async execute(logger, task, beginMessage, endMessage, errorMessage, errorHandler) {
     const uuid = uuidv1();