+++ tests/index.js
@@ -64,6 +64,36 @@ async function go(colors) {
   // =================
   // Ready for testing
 
+  // Test (Error, '')
+  {
+    logger.log(new Error('err'), 'message');
+    const entry = unitTest.entries[unitTest.console.entries.length - 2];
+    if (entry.level !== 'error') throw new Error();
+    if (!entry.message === 'message') throw new Error();
+    if (!entry.data.error) throw new Error();
+  }
+  {
+    logger.info(new Error('err'), 'message');
+    const entry = unitTest.entries[unitTest.console.entries.length - 2];
+    if (entry.level !== 'info') throw new Error();
+    if (!entry.message === 'message') throw new Error();
+    if (!entry.data.error) throw new Error();
+  }
+  {
+    logger.child().log(new Error('err'), 'message');
+    const entry = unitTest.entries[unitTest.console.entries.length - 2];
+    if (entry.level !== 'error') throw new Error();
+    if (!entry.message === 'message') throw new Error();
+    if (!entry.data.error) throw new Error();
+  }
+  {
+    logger.child().info(new Error('err'), 'message');
+    const entry = unitTest.entries[unitTest.console.entries.length - 2];
+    if (entry.level !== 'info') throw new Error();
+    if (!entry.message === 'message') throw new Error();
+    if (!entry.data.error) throw new Error();
+  }
+
   // Test passing category to logLevel
   {
     logger.info(['extra'], null, null, 'dragon');
@@ -90,6 +120,7 @@ async function go(colors) {
   {
     logger.error({ error: new Error('inner error'), message: { message: 'Foo', a: 5 } });
     const entry = unitTest.console.entries[unitTest.console.entries.length - 2];
+    if (entry.level !== 'error') throw new Error();
     if (!entry.data.error) throw new Error();
     if (!colors && entry.message !== 'Foo') throw new Error();
     if (!entry.data.a) throw new Error();
@@ -180,7 +211,7 @@ async function go(colors) {
 
   // This is logged as debug
   {
-    loggers.child('error').log(Loggers.tags({ logLevel: 'warn' }, { logLevel: 'debug' }), 'Yabba dabba');
+    loggers.child('error').log(loggers.tags({ logLevel: 'warn' }, { logLevel: 'debug' }), 'Yabba dabba');
     const { level } = unitTest.file.entries[unitTest.file.entries.length - 1];
     if (level !== 'debug') throw new Error();
   }
@@ -202,11 +233,11 @@ async function go(colors) {
     if (entry.message !== 'Error: 5') throw new Error();
   }
 
-  const tags = Loggers.tags('message');
+  const tags = loggers.tags('message');
   if (!tags.message) throw new Error();
 
   // Test disabling a tag
-  if (!loggers.isLevelEnabled(Loggers.tags({ silly: 1 }, { silly: 0 }))) throw new Error('isLevelEnabled failed');
+  if (!loggers.isLevelEnabled(loggers.tags({ silly: 1 }, { silly: 0 }))) throw new Error('isLevelEnabled failed');
 
   // Test isLevelEnabled
   if (!loggers.isLevelEnabled('debug')) throw new Error('isLevelEnabled failed');
@@ -221,7 +252,7 @@ async function go(colors) {
   });
 
   // Test 'ready'
-  if (!loggers.isReady()) throw new Error('ready failed');
+  if (!loggers.ready) throw new Error('ready failed');
 
   // Test two level names in tags
   {
@@ -622,7 +653,7 @@ async function go(colors) {
 
     setTimeout(() => {
       throw new Error('Unhandled exception');
-    });
+    }, .1);
     await new Promise((resolve) => setTimeout(resolve, 1));
 
     const len2 = Object.keys(unitTest.logGroupIds).length;
@@ -631,18 +662,18 @@ async function go(colors) {
 
   // ===============
   // Stop the logger
-  await new Promise((resolve) => setTimeout(() => loggers.stop().then(resolve), 100));
-
-  if (loggers.isReady()) throw new Error('ready failed');
+  if (!loggers.ready) throw new Error('ready failed');
+  await loggers.stop();
+  if (loggers.ready) throw new Error('ready failed');
 
   logger.info(`I've stopped and I can't get up`);
 
   {
     // These values must be tweaked whenever more entries are logged
-    if (unitTest.entries.length !== 146 + 10 * hasCloudWatch) throw new Error(unitTest.entries.length);
+    if (unitTest.entries.length !== 153 + 10 * hasCloudWatch) throw new Error(unitTest.entries.length);
     const len = Object.keys(unitTest.logGroupIds).length;
-    if (len !== 26) throw new Error(len);
-    if (unitTest.dataCount !== 91 + 10 * hasCloudWatch) throw new Error(unitTest.dataCount);
+    if (len !== 30) throw new Error(len);
+    if (unitTest.dataCount !== 98 + 10 * hasCloudWatch) throw new Error(unitTest.dataCount);
   }
 
   if (!onRan) throw new Error();